% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMM}
\alias{ADMM}
\title{ADMM}
\arguments{
\item{X}{matrix of independent  variables}

\item{Y}{vector of dependent variables}

\item{max_iter}{maximum number of iterations of the ADMM loop}

\item{rho}{augmented Lagrangian parameter}
}
\description{
ADMM to solve SLOPE-penalized ordinary least squares regression.
}
\examples{
library(SLOPE)
fit <- ADMM(as.matrix(bodyfat$x), bodyfat$y, 1e4, 1)

}
