\name{ADMMSLOPE-package}
\alias{ADMMSLOPE-package}
\alias{ADMMSLOPE}
\docType{package}
\title{
\packageTitle{ADMMSLOPE}
}
\description{
\packageDescription{ADMMSLOPE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ADMMSLOPE}
\packageIndices{ADMMSLOPE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ADMMSLOPE}

Maintainer: \packageMaintainer{ADMMSLOPE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
