% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastProxSL1}
\alias{FastProxSL1}
\title{FastProxSL1}
\arguments{
\item{beta}{input vector}

\item{lambda}{vector of lambdas, sorted in decreasing order}
}
\description{
A function using RcppArmadillo that computes the proximal operator for SLOPE using Algorithm 3 (FastProxSL1) from Bogdan et al 2015 (SLOPE: adaptive variable selection via convex optimization).
}
\examples{
p = 1e7
}
